Setup Auth0 and guard CMS
