+++ app/api/auth/[auth0]/route.ts
@@ -1,0 +1,3 @@
+import Auth0 from "#/lib/auth0";
+
+export const GET = Auth0.handleAuth();
