+++ lib/auth0.ts
@@ -1,0 +1,9 @@
+import { initAuth0 } from "@auth0/nextjs-auth0";
+
+export default initAuth0({
+  secret: process.env.AUTH0_SECRET,
+  issuerBaseURL: process.env.AUTH0_ISSUER_BASE_URL,
+  baseURL: process.env.VERCEL_URL ? `https://${process.env.VERCEL_URL}` : "http://localhost:3000",
+  clientID: process.env.AUTH0_CLIENT_ID,
+  clientSecret: process.env.AUTH0_CLIENT_SECRET,
+});
