+++ package.json
@@ -18,6 +18,7 @@
     "test:chromatic": "npx chromatic --only-changed --build-script-name=build:storybook"
   },
   "dependencies": {
+    "@auth0/nextjs-auth0": "^3.5.0",
     "@next/mdx": "^14.1.0",
     "@react-pdf-viewer/core": "^3.12.0",
     "@react-pdf-viewer/default-layout": "^3.12.0",
